Rename only - LinkConnectorEventMap
