+++ src/canvas/LinkConnector.ts
@@ -7,7 +7,7 @@ import type { CanvasPointerEvent } from "@/types/events"
 import type { IWidget } from "@/types/widgets"
 
 import { CustomEventTarget } from "@/infrastructure/CustomEventTarget"
-import { LinkConnectorEventMap } from "@/infrastructure/LinkConnectorEventTarget"
+import { LinkConnectorEventMap } from "@/infrastructure/LinkConnectorEventMap"
 import { LLink } from "@/LLink"
 import { LinkDirection } from "@/types/globalEnums"
